Make proxy abstract.
